corrected difusioLT.cpp, changes to main methods of difusioLT and difusioIC
there is a little bug in the update iteration of the Linear Model.

since the active vector is being updated during the iteration over all nodes, the iterations might cross influence each other.

I changed this to work in the way in the way it was described in the pdf.

First it is checked for all nodes if there are enough active neighbors to influence their activation.

Only after this is checked for all neighbors the update is made (if there is any).

___________________________

Additionally I updated the main methods of the two difusion models to allow to pass arguments p or r.